Core/Common: Merge TimeTrackerSmall with TimeTracker
Port From (https://github.com/TrinityCore/TrinityCore/commit/8809d54ca2daa230b1439f012c8b10ee88d6beae)

+++ Source/Game/Entities/Player/Player.cs
@@ -612,7 +612,7 @@ namespace Game.Entities
             }
 
             // group update
-            m_groupUpdateTimer.Update((int)diff);
+            m_groupUpdateTimer.Update(diff);
             if (m_groupUpdateTimer.Passed())
             {
                 SendUpdateToOutOfRangeGroupMembers();
